+++ xbmc/filesystem/FileReaderFile.cpp
@@ -67,7 +67,7 @@ bool CFileReaderFile::OpenForWrite(const CURL& url, bool bOverWrite)
 }
 
 //*********************************************************************************************
-unsigned int CFileReaderFile::Read(void *lpBuf, int64_t uiBufSize)
+ssize_t CFileReaderFile::Read(void *lpBuf, size_t uiBufSize)
 {
   return m_reader.Read(lpBuf,uiBufSize);
 }
